find gcloud_bin_path executable (resolving symlink) using realpath and whence, Fixes #9
`whence` is a zsh builtin and is more reliable than using `which` (beacuse it may be either an executaable such as /usr/bin/which and is also a zsh builtin).

When used as `whence -p` it will ignore shell functions and aliases, instead searching the path for the executable.

From the zsh builtins manpage:
              -p     Do a path search for name even if it is an alias, reserved word, shell function or builtin.

Fixes #9: gcloud docs location was not in same direction as "which gcloud" supplied.